[i18n] HashMAC fix, Hash-Collision fix -- intermediate state]

+++ CrypTool/CryptDoc.cpp
@@ -1406,7 +1406,7 @@ void CCryptDoc::OnAnalyseZufallstestsFipspub1401()
 	file.Open(ContentName, CFile::modeRead);
 	if(file.GetLength() < 2500) 
 	{
-		str_size_fmt = itoa_fmt(file.GetLength());
+		str_size_fmt = itoa_fmt((__int64)file.GetLength());
 		Message(IDS_STRING_TEST_FIPS_KLEINE_DATEI_2, MB_ICONEXCLAMATION, str_size_fmt);
 		delete []str_size_fmt;
 		file.Close();
@@ -1415,7 +1415,7 @@ void CCryptDoc::OnAnalyseZufallstestsFipspub1401()
 	{
 		if (file.GetLength() > 2500)
 		{
-			str_size_fmt = itoa_fmt((unsigned long)file.GetLength());
+			str_size_fmt = itoa_fmt((int)file.GetLength());
 			Message(IDS_STRING_TEST_FIPS_KLEINE_DATEI, MB_ICONINFORMATION, str_size_fmt);
 			delete []str_size_fmt;
 		}
